// frontend/src/api-client/models/CreateClientDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VET API
 * CRM pour vétérinaires
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateClientDto
 */
export interface CreateClientDto {
    /**
     * 
     * @type {string}
     * @memberof CreateClientDto
     */
    civility: CreateClientDtoCivilityEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateClientDto
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof CreateClientDto
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof CreateClientDto
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof CreateClientDto
     */
    phone: string;
}


/**
 * @export
 */
export const CreateClientDtoCivilityEnum = {
    M: 'M.',
    Mme: 'Mme',
    Mlle: 'Mlle'
} as const;
export type CreateClientDtoCivilityEnum = typeof CreateClientDtoCivilityEnum[keyof typeof CreateClientDtoCivilityEnum];


/**
 * Check if a given object implements the CreateClientDto interface.
 */
export function instanceOfCreateClientDto(value: object): boolean {
    if (!('civility' in value)) return false;
    if (!('firstName' in value)) return false;
    if (!('lastName' in value)) return false;
    if (!('email' in value)) return false;
    if (!('phone' in value)) return false;
    return true;
}

export function CreateClientDtoFromJSON(json: any): CreateClientDto {
    return CreateClientDtoFromJSONTyped(json, false);
}

export function CreateClientDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateClientDto {
    if (json == null) {
        return json;
    }
    return {
        
        'civility': json['civility'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'email': json['email'],
        'phone': json['phone'],
    };
}

export function CreateClientDtoToJSON(value?: CreateClientDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'civility': value['civility'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'email': value['email'],
        'phone': value['phone'],
    };
}