// frontend/src/api-client/models/UpdateClientDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VET API
 * CRM pour vétérinaires
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateClientDto
 */
export interface UpdateClientDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateClientDto
     */
    civility?: UpdateClientDtoCivilityEnum;
    /**
     * 
     * @type {string}
     * @memberof UpdateClientDto
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateClientDto
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateClientDto
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateClientDto
     */
    phone?: string;
}


/**
 * @export
 */
export const UpdateClientDtoCivilityEnum = {
    M: 'M.',
    Mme: 'Mme',
    Mlle: 'Mlle'
} as const;
export type UpdateClientDtoCivilityEnum = typeof UpdateClientDtoCivilityEnum[keyof typeof UpdateClientDtoCivilityEnum];


/**
 * Check if a given object implements the UpdateClientDto interface.
 */
export function instanceOfUpdateClientDto(value: object): boolean {
    return true;
}

export function UpdateClientDtoFromJSON(json: any): UpdateClientDto {
    return UpdateClientDtoFromJSONTyped(json, false);
}

export function UpdateClientDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateClientDto {
    if (json == null) {
        return json;
    }
    return {
        
        'civility': json['civility'] == null ? undefined : json['civility'],
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'email': json['email'] == null ? undefined : json['email'],
        'phone': json['phone'] == null ? undefined : json['phone'],
    };
}

export function UpdateClientDtoToJSON(value?: UpdateClientDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'civility': value['civility'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'email': value['email'],
        'phone': value['phone'],
    };
}