// frontend/api-client/models/Client.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VET API
 * CRM pour vétérinaires
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Animal } from './Animal';
import {
    AnimalFromJSON,
    AnimalFromJSONTyped,
    AnimalToJSON,
} from './Animal';

/**
 * 
 * @export
 * @interface Client
 */
export interface Client {
    /**
     * 
     * @type {number}
     * @memberof Client
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    civility: ClientCivilityEnum;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof Client
     */
    phone: string;
    /**
     * 
     * @type {Array<Animal>}
     * @memberof Client
     */
    animals?: Array<Animal>;
    /**
     * 
     * @type {Date}
     * @memberof Client
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Client
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const ClientCivilityEnum = {
    M: 'M.',
    Mme: 'Mme',
    Mlle: 'Mlle'
} as const;
export type ClientCivilityEnum = typeof ClientCivilityEnum[keyof typeof ClientCivilityEnum];


/**
 * Check if a given object implements the Client interface.
 */
export function instanceOfClient(value: object): boolean {
    if (!('id' in value)) return false;
    if (!('civility' in value)) return false;
    if (!('firstName' in value)) return false;
    if (!('lastName' in value)) return false;
    if (!('email' in value)) return false;
    if (!('phone' in value)) return false;
    if (!('createdAt' in value)) return false;
    if (!('updatedAt' in value)) return false;
    return true;
}

export function ClientFromJSON(json: any): Client {
    return ClientFromJSONTyped(json, false);
}

export function ClientFromJSONTyped(json: any, ignoreDiscriminator: boolean): Client {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'civility': json['civility'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'email': json['email'],
        'phone': json['phone'],
        'animals': json['animals'] == null ? undefined : ((json['animals'] as Array<any>).map(AnimalFromJSON)),
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function ClientToJSON(value?: Client | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'civility': value['civility'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'email': value['email'],
        'phone': value['phone'],
        'animals': value['animals'] == null ? undefined : ((value['animals'] as Array<any>).map(AnimalToJSON)),
        'createdAt': ((value['createdAt']).toISOString()),
        'updatedAt': ((value['updatedAt']).toISOString()),
    };
}