// frontend/api-client/models/CreateAnimalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VET API
 * CRM pour vétérinaires
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateAnimalDto
 */
export interface CreateAnimalDto {
    /**
     * 
     * @type {string}
     * @memberof CreateAnimalDto
     */
    firstName: string;
    /**
     * 
     * @type {number}
     * @memberof CreateAnimalDto
     */
    age: number;
    /**
     * 
     * @type {number}
     * @memberof CreateAnimalDto
     */
    height: number;
    /**
     * 
     * @type {number}
     * @memberof CreateAnimalDto
     */
    weight: number;
    /**
     * 
     * @type {string}
     * @memberof CreateAnimalDto
     */
    species: CreateAnimalDtoSpeciesEnum;
    /**
     * 
     * @type {number}
     * @memberof CreateAnimalDto
     */
    clientId: number;
}


/**
 * @export
 */
export const CreateAnimalDtoSpeciesEnum = {
    Chien: 'chien',
    Chat: 'chat',
    Hamster: 'hamster',
    Tortue: 'tortue'
} as const;
export type CreateAnimalDtoSpeciesEnum = typeof CreateAnimalDtoSpeciesEnum[keyof typeof CreateAnimalDtoSpeciesEnum];


/**
 * Check if a given object implements the CreateAnimalDto interface.
 */
export function instanceOfCreateAnimalDto(value: object): boolean {
    if (!('firstName' in value)) return false;
    if (!('age' in value)) return false;
    if (!('height' in value)) return false;
    if (!('weight' in value)) return false;
    if (!('species' in value)) return false;
    if (!('clientId' in value)) return false;
    return true;
}

export function CreateAnimalDtoFromJSON(json: any): CreateAnimalDto {
    return CreateAnimalDtoFromJSONTyped(json, false);
}

export function CreateAnimalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateAnimalDto {
    if (json == null) {
        return json;
    }
    return {
        
        'firstName': json['firstName'],
        'age': json['age'],
        'height': json['height'],
        'weight': json['weight'],
        'species': json['species'],
        'clientId': json['clientId'],
    };
}

export function CreateAnimalDtoToJSON(value?: CreateAnimalDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'firstName': value['firstName'],
        'age': value['age'],
        'height': value['height'],
        'weight': value['weight'],
        'species': value['species'],
        'clientId': value['clientId'],
    };
}