// frontend/api-client/models/UpdateAnimalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VET API
 * CRM pour vétérinaires
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateAnimalDto
 */
export interface UpdateAnimalDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateAnimalDto
     */
    firstName?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateAnimalDto
     */
    age?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateAnimalDto
     */
    height?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateAnimalDto
     */
    weight?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateAnimalDto
     */
    species?: UpdateAnimalDtoSpeciesEnum;
    /**
     * 
     * @type {number}
     * @memberof UpdateAnimalDto
     */
    clientId?: number;
}


/**
 * @export
 */
export const UpdateAnimalDtoSpeciesEnum = {
    Chien: 'chien',
    Chat: 'chat',
    Hamster: 'hamster',
    Tortue: 'tortue'
} as const;
export type UpdateAnimalDtoSpeciesEnum = typeof UpdateAnimalDtoSpeciesEnum[keyof typeof UpdateAnimalDtoSpeciesEnum];


/**
 * Check if a given object implements the UpdateAnimalDto interface.
 */
export function instanceOfUpdateAnimalDto(value: object): boolean {
    return true;
}

export function UpdateAnimalDtoFromJSON(json: any): UpdateAnimalDto {
    return UpdateAnimalDtoFromJSONTyped(json, false);
}

export function UpdateAnimalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateAnimalDto {
    if (json == null) {
        return json;
    }
    return {
        
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
        'age': json['age'] == null ? undefined : json['age'],
        'height': json['height'] == null ? undefined : json['height'],
        'weight': json['weight'] == null ? undefined : json['weight'],
        'species': json['species'] == null ? undefined : json['species'],
        'clientId': json['clientId'] == null ? undefined : json['clientId'],
    };
}

export function UpdateAnimalDtoToJSON(value?: UpdateAnimalDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'firstName': value['firstName'],
        'age': value['age'],
        'height': value['height'],
        'weight': value['weight'],
        'species': value['species'],
        'clientId': value['clientId'],
    };
}